+++ CommonAssemblyInfo.cs
@@ -49,9 +49,9 @@ using System.Reflection;
 //Release Candidate: 1.8.0.3001 (Displayed as "1.8 RC1")
 //Beta: 1.8.0.2004 (Displayed as "1.8 BETA004")
 //Develop: 1.8.0.1022 (Displayed as "1.8 NIGHTLY #022")
-[assembly: AssemblyVersion("3.0.0.2017")]
-[assembly: AssemblyFileVersion("3.0.0.2017")]
-[assembly: AssemblyInformationalVersion("3.0.0.2017-beta")]
+[assembly: AssemblyVersion("3.0.0.2019")]
+[assembly: AssemblyFileVersion("3.0.0.2019")]
+[assembly: AssemblyInformationalVersion("3.0.0.2019-beta")]
 
 // The PluginMinimumApplicationVersion specifies the minimum application version the plugin needs to specify in order to be compatible
 // Plugins that have a version specified that is below this version are marked as incompatible